src_vanna/requirements.txt
#pysqlite3-binary
chromadb
vanna[chromadb,openai]
openai
azure-identity
azure-ai-ml
flask
pyodbc
db-dtypes
python-dotenv
gunicorn